fix load const

+++ src/bytecode.rs
@@ -4,7 +4,7 @@ pub enum ByteCode {
     SetGlobal(u8, u8),
     SetGlobalConst(u8, u8),
     SetGlobalGlobal(u8, u8),
-    LoadConst(u8, u8),
+    LoadConst(u8, u16),
     LoadNil(u8),
     LoadBool(u8, bool),
     LoadInt(u8, i16),
