+++ src/parse.rs
@@ -161,7 +161,7 @@ impl ParseProto {
     }
 
     fn load_const(&mut self, dst: usize, val: Value) -> ByteCode {
-        ByteCode::LoadConst(dst as u8, self.add_const(val) as u8)
+        ByteCode::LoadConst(dst as u8, self.add_const(val) as u16)
     }
 
     fn load_var(&mut self, dst: usize, name: String) -> ByteCode {
